ocr_service/requirements.txt
pytest==8.2.2
pylint==3.2.5
black==24.4.2
nbval==0.11.0
pytest-cov==5.0.0
ipython==8.26.0
pillow==10.4.0
pytesseract==0.3.10
fire==0.6.0
python-multipart==0.0.9
fastapi==0.111.0